Refactor the relationship between the Tagger and Notifier.
